beginnings of a refactor
