+++ libs/langchain_v1/langchain/agents/__init__.py
@@ -1,6 +1,7 @@
 """langgraph.prebuilt exposes a higher-level API for creating and executing agents and tools."""
 
-from langchain.agents.react_agent import AgentState, create_agent
+from langchain.agents.factory import create_agent
+from langchain.agents.middleware.types import AgentState
 from langchain.tools import ToolNode
 
 __all__ = [
