+++ libs/langchain_v1/tests/unit_tests/agents/test_middleware_agent.py
@@ -51,7 +51,7 @@ from langchain.agents.middleware.types import (
     OmitFromOutput,
     PrivateStateAttr,
 )
-from langchain.agents.middleware_agent import create_agent
+from langchain.agents.factory import create_agent
 from langchain.agents.structured_output import ToolStrategy
 from langchain.tools import InjectedState
 
@@ -133,7 +133,7 @@ def test_create_agent_diagram(
         system_prompt="You are a helpful assistant.",
     )
 
-    assert agent_zero.compile().get_graph().draw_mermaid() == snapshot
+    assert agent_zero.get_graph().draw_mermaid() == snapshot
 
     agent_one = create_agent(
         model=FakeToolCallingModel(),
@@ -142,7 +142,7 @@ def test_create_agent_diagram(
         middleware=[NoopOne()],
     )
 
-    assert agent_one.compile().get_graph().draw_mermaid() == snapshot
+    assert agent_one.get_graph().draw_mermaid() == snapshot
 
     agent_two = create_agent(
         model=FakeToolCallingModel(),
@@ -151,7 +151,7 @@ def test_create_agent_diagram(
         middleware=[NoopOne(), NoopTwo()],
     )
 
-    assert agent_two.compile().get_graph().draw_mermaid() == snapshot
+    assert agent_two.get_graph().draw_mermaid() == snapshot
 
     agent_three = create_agent(
         model=FakeToolCallingModel(),
@@ -160,7 +160,7 @@ def test_create_agent_diagram(
         middleware=[NoopOne(), NoopTwo(), NoopThree()],
     )
 
-    assert agent_three.compile().get_graph().draw_mermaid() == snapshot
+    assert agent_three.get_graph().draw_mermaid() == snapshot
 
     agent_four = create_agent(
         model=FakeToolCallingModel(),
@@ -169,7 +169,7 @@ def test_create_agent_diagram(
         middleware=[NoopFour()],
     )
 
-    assert agent_four.compile().get_graph().draw_mermaid() == snapshot
+    assert agent_four.get_graph().draw_mermaid() == snapshot
 
     agent_five = create_agent(
         model=FakeToolCallingModel(),
@@ -178,7 +178,7 @@ def test_create_agent_diagram(
         middleware=[NoopFour(), NoopFive()],
     )
 
-    assert agent_five.compile().get_graph().draw_mermaid() == snapshot
+    assert agent_five.get_graph().draw_mermaid() == snapshot
 
     agent_six = create_agent(
         model=FakeToolCallingModel(),
@@ -187,7 +187,7 @@ def test_create_agent_diagram(
         middleware=[NoopFour(), NoopFive(), NoopSix()],
     )
 
-    assert agent_six.compile().get_graph().draw_mermaid() == snapshot
+    assert agent_six.get_graph().draw_mermaid() == snapshot
 
     agent_seven = create_agent(
         model=FakeToolCallingModel(),
@@ -196,7 +196,7 @@ def test_create_agent_diagram(
         middleware=[NoopSeven()],
     )
 
-    assert agent_seven.compile().get_graph().draw_mermaid() == snapshot
+    assert agent_seven.get_graph().draw_mermaid() == snapshot
 
     agent_eight = create_agent(
         model=FakeToolCallingModel(),
@@ -205,7 +205,7 @@ def test_create_agent_diagram(
         middleware=[NoopSeven(), NoopEight()],
     )
 
-    assert agent_eight.compile().get_graph().draw_mermaid() == snapshot
+    assert agent_eight.get_graph().draw_mermaid() == snapshot
 
     agent_nine = create_agent(
         model=FakeToolCallingModel(),
@@ -214,7 +214,7 @@ def test_create_agent_diagram(
         middleware=[NoopSeven(), NoopEight(), NoopNine()],
     )
 
-    assert agent_nine.compile().get_graph().draw_mermaid() == snapshot
+    assert agent_nine.get_graph().draw_mermaid() == snapshot
 
     agent_ten = create_agent(
         model=FakeToolCallingModel(),
@@ -223,7 +223,7 @@ def test_create_agent_diagram(
         middleware=[NoopTen()],
     )
 
-    assert agent_ten.compile().get_graph().draw_mermaid() == snapshot
+    assert agent_ten.get_graph().draw_mermaid() == snapshot
 
     agent_eleven = create_agent(
         model=FakeToolCallingModel(),
@@ -232,7 +232,7 @@ def test_create_agent_diagram(
         middleware=[NoopTen(), NoopEleven()],
     )
 
-    assert agent_eleven.compile().get_graph().draw_mermaid() == snapshot
+    assert agent_eleven.get_graph().draw_mermaid() == snapshot
 
 
 def test_create_agent_invoke(
@@ -281,7 +281,8 @@ def test_create_agent_invoke(
         tools=[my_tool],
         system_prompt="You are a helpful assistant.",
         middleware=[NoopSeven(), NoopEight()],
-    ).compile(checkpointer=sync_checkpointer)
+        checkpointer=sync_checkpointer,
+    )
 
     thread1 = {"configurable": {"thread_id": "1"}}
     assert agent_one.invoke({"messages": ["hello"]}, thread1) == {
@@ -370,7 +371,8 @@ def test_create_agent_jump(
         tools=[my_tool],
         system_prompt="You are a helpful assistant.",
         middleware=[NoopSeven(), NoopEight()],
-    ).compile(checkpointer=sync_checkpointer)
+        checkpointer=sync_checkpointer,
+    )
 
     if isinstance(sync_checkpointer, InMemorySaver):
         assert agent_one.get_graph().draw_mermaid() == snapshot
@@ -1214,14 +1216,13 @@ def test_modify_model_request() -> None:
             request.messages.append(HumanMessage("remember to be nice!"))
             return request
 
-    builder = create_agent(
+    agent = create_agent(
         model=FakeToolCallingModel(),
         tools=[],
         system_prompt="You are a helpful assistant.",
         middleware=[ModifyMiddleware()],
     )
 
-    agent = builder.compile()
     result = agent.invoke({"messages": [HumanMessage("Hello")]})
     assert result["messages"][0].content == "Hello"
     assert result["messages"][1].content == "remember to be nice!"
@@ -1268,9 +1269,8 @@ def test_tools_to_model_edge_with_structured_and_regular_tool_calls():
         response_format=ToolStrategy(schema=WeatherResponse),
     )
 
-    # Compile and invoke the agent
-    compiled_agent = agent.compile()
-    result = compiled_agent.invoke(
+    # Invoke the agent
+    result = agent.invoke(
         {"messages": [HumanMessage("What's the weather and help me with a query?")]}
     )
 
@@ -1323,7 +1323,7 @@ def test_public_private_state_for_custom_middleware() -> None:
             return {"omit_input": "test", "omit_output": "test", "private_state": "test"}
 
     agent = create_agent(model=FakeToolCallingModel(), middleware=[CustomMiddleware()])
-    agent = agent.compile()
+    # agent is already compiled
     result = agent.invoke(
         {
             "messages": [HumanMessage("Hello")],
@@ -1358,7 +1358,7 @@ def test_runtime_injected_into_middleware() -> None:
     middleware = CustomMiddleware()
 
     agent = create_agent(model=FakeToolCallingModel(), middleware=[CustomMiddleware()])
-    agent = agent.compile()
+    # agent is already compiled
     agent.invoke({"messages": [HumanMessage("Hello")]})
 
 
@@ -1389,7 +1389,7 @@ def test_injected_state_in_middleware_agent() -> None:
         tools=[test_state],
         system_prompt="You are a helpful assistant.",
         middleware=[TestMiddleware()],
-    ).compile()
+    )
 
     result = agent.invoke(
         {"test_state": "I love pizza", "messages": [HumanMessage("Call the test state tool")]}
@@ -1419,7 +1419,7 @@ def test_jump_to_is_ephemeral() -> None:
             return {"jump_to": "model"}
 
     agent = create_agent(model=FakeToolCallingModel(), middleware=[MyMiddleware()])
-    agent = agent.compile()
+    # agent is already compiled
     result = agent.invoke({"messages": [HumanMessage("Hello")]})
     assert "jump_to" not in result
 
@@ -1550,7 +1550,7 @@ def test_planning_middleware_agent_creation_with_middleware() -> None:
     )
     middleware = PlanningMiddleware()
     agent = create_agent(model=model, middleware=[middleware])
-    agent = agent.compile()
+    # agent is already compiled
 
     result = agent.invoke({"messages": [HumanMessage("Hello")]})
     assert result["todos"] == [{"content": "Task 1", "status": "completed"}]
@@ -1661,7 +1661,7 @@ def test_planning_middleware_custom_system_prompt() -> None:
     )
 
     agent = create_agent(model=model, middleware=[middleware])
-    agent = agent.compile()
+    # agent is already compiled
 
     result = agent.invoke({"messages": [HumanMessage("Hello")]})
     assert result["todos"] == [{"content": "Custom task", "status": "pending"}]
@@ -1735,7 +1735,8 @@ def test_thread_limit_with_create_agent():
         model=model,
         tools=[simple_tool],
         middleware=[ModelCallLimitMiddleware(thread_limit=1)],
-    ).compile(checkpointer=InMemorySaver())
+        checkpointer=InMemorySaver(),
+    )
 
     # First invocation should work - 1 model call, within thread limit
     result = agent.invoke(
@@ -1778,7 +1779,8 @@ def test_run_limit_with_create_agent():
         model=model,
         tools=[simple_tool],
         middleware=[ModelCallLimitMiddleware(run_limit=1)],
-    ).compile(checkpointer=InMemorySaver())
+        checkpointer=InMemorySaver(),
+    )
 
     # This should hit the run limit after the first model call
     result = agent.invoke(
@@ -1865,8 +1867,7 @@ def test_run_limit_resets_between_invocations() -> None:
         tool_calls=[[], [], [], []]
     )  # No tool calls, so only model call per run
 
-    agent = create_agent(model=model, middleware=[middleware])
-    agent = agent.compile(checkpointer=InMemorySaver())
+    agent = create_agent(model=model, middleware=[middleware], checkpointer=InMemorySaver())
 
     thread_config = {"configurable": {"thread_id": "test_thread"}}
     agent.invoke({"messages": [HumanMessage("Hello")]}, thread_config)
@@ -1913,7 +1914,7 @@ async def test_create_agent_async_invoke() -> None:
         tools=[my_tool],
         system_prompt="You are a helpful assistant.",
         middleware=[AsyncMiddleware()],
-    ).compile()
+    )
 
     result = await agent.ainvoke({"messages": [HumanMessage("hello")]})
 
@@ -1969,7 +1970,7 @@ async def test_create_agent_async_invoke_multiple_middleware() -> None:
         tools=[],
         system_prompt="You are a helpful assistant.",
         middleware=[AsyncMiddlewareOne(), AsyncMiddlewareTwo()],
-    ).compile()
+    )
 
     result = await agent.ainvoke({"messages": [HumanMessage("hello")]})
 
@@ -2010,7 +2011,7 @@ async def test_create_agent_async_jump() -> None:
         tools=[my_tool],
         system_prompt="You are a helpful assistant.",
         middleware=[AsyncMiddlewareOne(), AsyncMiddlewareTwo()],
-    ).compile()
+    )
 
     result = await agent.ainvoke({"messages": []})
 
@@ -2049,7 +2050,7 @@ async def test_create_agent_mixed_sync_async_middleware() -> None:
         tools=[],
         system_prompt="You are a helpful assistant.",
         middleware=[SyncMiddleware(), AsyncMiddleware()],
-    ).compile()
+    )
 
     result = await agent.ainvoke({"messages": [HumanMessage("hello")]})
 
@@ -2076,7 +2077,7 @@ def test_create_agent_sync_invoke_with_only_async_middleware_raises_error() -> N
         tools=[],
         system_prompt="You are a helpful assistant.",
         middleware=[AsyncOnlyMiddleware()],
-    ).compile()
+    )
 
     with pytest.raises(
         TypeError,
@@ -2109,7 +2110,7 @@ def test_create_agent_sync_invoke_with_mixed_middleware() -> None:
         tools=[],
         system_prompt="You are a helpful assistant.",
         middleware=[MixedMiddleware()],
-    ).compile()
+    )
 
     result = agent.invoke({"messages": [HumanMessage("hello")]})
