+++ libs/langchain_v1/tests/unit_tests/agents/test_middleware_decorators.py
@@ -19,7 +19,7 @@ from langchain.agents.middleware.types import (
     modify_model_request,
     hook_config,
 )
-from langchain.agents.middleware_agent import create_agent, _get_can_jump_to
+from langchain.agents.factory import create_agent, _get_can_jump_to
 from .model import FakeToolCallingModel
 
 
@@ -136,7 +136,7 @@ def test_all_decorators_integration() -> None:
     agent = create_agent(
         model=FakeToolCallingModel(), middleware=[track_before, track_modify, track_after]
     )
-    agent = agent.compile()
+    # agent is already compiled
     agent.invoke({"messages": [HumanMessage("Hello")]})
 
     assert call_order == ["before", "modify", "after"]
@@ -225,7 +225,7 @@ def test_can_jump_to_integration() -> None:
         return None
 
     agent = create_agent(model=FakeToolCallingModel(), middleware=[early_exit])
-    agent = agent.compile()
+    # agent is already compiled
 
     # Test with early exit
     result = agent.invoke({"messages": [HumanMessage("exit")]})
@@ -338,7 +338,7 @@ async def test_async_decorators_integration() -> None:
         model=FakeToolCallingModel(),
         middleware=[track_async_before, track_async_modify, track_async_after],
     )
-    agent = agent.compile()
+    # agent is already compiled
     await agent.ainvoke({"messages": [HumanMessage("Hello")]})
 
     assert call_order == ["async_before", "async_modify", "async_after"]
@@ -394,7 +394,7 @@ async def test_mixed_sync_async_decorators_integration() -> None:
             track_sync_after,
         ],
     )
-    agent = agent.compile()
+    # agent is already compiled
     await agent.ainvoke({"messages": [HumanMessage("Hello")]})
 
     assert call_order == [
@@ -455,7 +455,7 @@ async def test_async_can_jump_to_integration() -> None:
         return None
 
     agent = create_agent(model=FakeToolCallingModel(), middleware=[async_early_exit])
-    agent = agent.compile()
+    # agent is already compiled
 
     # Test with early exit
     result = await agent.ainvoke({"messages": [HumanMessage("exit")]})
@@ -526,7 +526,7 @@ def test_async_middleware_with_can_jump_to_graph_snapshot(snapshot: SnapshotAsse
         model=FakeToolCallingModel(), middleware=[async_before_with_jump]
     )
 
-    assert agent_async_before.compile().get_graph().draw_mermaid() == snapshot
+    assert agent_async_before.get_graph().draw_mermaid() == snapshot
 
     # Test 2: Async after_model with can_jump_to
     @after_model(can_jump_to=["model", "end"])
@@ -539,7 +539,7 @@ def test_async_middleware_with_can_jump_to_graph_snapshot(snapshot: SnapshotAsse
         model=FakeToolCallingModel(), middleware=[async_after_with_jump]
     )
 
-    assert agent_async_after.compile().get_graph().draw_mermaid() == snapshot
+    assert agent_async_after.get_graph().draw_mermaid() == snapshot
 
     # Test 3: Multiple async middleware with can_jump_to
     @before_model(can_jump_to=["end"])
@@ -555,7 +555,7 @@ def test_async_middleware_with_can_jump_to_graph_snapshot(snapshot: SnapshotAsse
         middleware=[async_before_early_exit, async_after_retry],
     )
 
-    assert agent_multiple_async.compile().get_graph().draw_mermaid() == snapshot
+    assert agent_multiple_async.get_graph().draw_mermaid() == snapshot
 
     # Test 4: Mixed sync and async middleware with can_jump_to
     @before_model(can_jump_to=["end"])
@@ -571,4 +571,4 @@ def test_async_middleware_with_can_jump_to_graph_snapshot(snapshot: SnapshotAsse
         middleware=[sync_before_with_jump, async_after_with_jumps],
     )
 
-    assert agent_mixed.compile().get_graph().draw_mermaid() == snapshot
+    assert agent_mixed.get_graph().draw_mermaid() == snapshot
