+++ libs/langchain_v1/tests/unit_tests/agents/test_middleware_tools.py
@@ -3,7 +3,7 @@
 import pytest
 
 from langchain.agents.middleware.types import AgentMiddleware, AgentState, ModelRequest
-from langchain.agents.middleware_agent import create_agent
+from langchain.agents.factory import create_agent
 from langchain_core.messages import HumanMessage, ToolMessage
 from langchain_core.tools import tool
 from .model import FakeToolCallingModel
@@ -36,7 +36,7 @@ def test_model_request_tools_are_strings() -> None:
         tools=[search_tool, calculator],
         system_prompt="You are a helpful assistant.",
         middleware=[RequestCapturingMiddleware()],
-    ).compile()
+    )
 
     agent.invoke({"messages": [HumanMessage("Hello")]})
 
@@ -87,7 +87,7 @@ def test_middleware_can_modify_tool_names() -> None:
         tools=[tool_a, tool_b, tool_c],
         system_prompt="You are a helpful assistant.",
         middleware=[ToolFilteringMiddleware()],
-    ).compile()
+    )
 
     result = agent.invoke({"messages": [HumanMessage("Use tool_a")]})
 
@@ -119,7 +119,7 @@ def test_unknown_tool_name_raises_error() -> None:
         tools=[known_tool],
         system_prompt="You are a helpful assistant.",
         middleware=[BadMiddleware()],
-    ).compile()
+    )
 
     with pytest.raises(ValueError, match="Middleware returned unknown tool names"):
         agent.invoke({"messages": [HumanMessage("Hello")]})
@@ -159,7 +159,7 @@ def test_middleware_can_add_and_remove_tools() -> None:
         tools=[search, admin_tool],
         system_prompt="You are a helpful assistant.",
         middleware=[ConditionalToolMiddleware()],
-    ).compile()
+    )
 
     # Test non-admin user - should not have access to admin_tool
     # We can't directly inspect the bound model, but we can verify the agent runs
@@ -194,7 +194,7 @@ def test_empty_tools_list_is_valid() -> None:
         tools=[some_tool],
         system_prompt="You are a helpful assistant.",
         middleware=[NoToolsMiddleware()],
-    ).compile()
+    )
 
     # Should run without error even with no tools
     result = agent.invoke({"messages": [HumanMessage("Hello")]})
@@ -245,7 +245,7 @@ def test_tools_preserved_across_multiple_middleware() -> None:
         tools=[tool_a, tool_b, tool_c],
         system_prompt="You are a helpful assistant.",
         middleware=[FirstMiddleware(), SecondMiddleware()],
-    ).compile()
+    )
 
     agent.invoke({"messages": [HumanMessage("Hello")]})
 
@@ -286,7 +286,7 @@ def test_middleware_with_additional_tools() -> None:
         tools=[base_tool],
         system_prompt="You are a helpful assistant.",
         middleware=[ToolProvidingMiddleware()],
-    ).compile()
+    )
 
     result = agent.invoke({"messages": [HumanMessage("Use middleware tool")]})
